info
One Piece - A pirate group with dreams of achieving their goals.
Black Adam - A hero in the movie but villain in the comics.
Chainsaw Man - A teen possessed by a devil and becomes a hybrid devil. Using chainsaw to kill other devils.
Spy x Family - A family with Spy, Assassin, Telepathy, and A dog that can see future.